Refactor - renamed app_logs to variant_logs
